refactor: Auction 레이어 기본 구조 수정
