+++ backend/src/main/java/com/sjiwon/anotherart/art/domain/service/ArtDeleter.java
@@ -3,7 +3,7 @@ package com.sjiwon.anotherart.art.domain.service;
 import com.sjiwon.anotherart.art.domain.model.Art;
 import com.sjiwon.anotherart.art.domain.repository.ArtRepository;
 import com.sjiwon.anotherart.art.domain.repository.HashtagRepository;
-import com.sjiwon.anotherart.auction.domain.AuctionRepository;
+import com.sjiwon.anotherart.auction.domain.repository.AuctionRepository;
 import com.sjiwon.anotherart.favorite.domain.repository.FavoriteRepository;
 import com.sjiwon.anotherart.global.annotation.AnotherArtWritableTransactional;
 import lombok.RequiredArgsConstructor;
