+++ backend/src/main/java/com/sjiwon/anotherart/art/domain/service/ArtDeletionPreInspector.java
@@ -2,7 +2,7 @@ package com.sjiwon.anotherart.art.domain.service;
 
 import com.sjiwon.anotherart.art.domain.model.Art;
 import com.sjiwon.anotherart.art.exception.ArtErrorCode;
-import com.sjiwon.anotherart.auction.domain.AuctionRepository;
+import com.sjiwon.anotherart.auction.domain.repository.AuctionRepository;
 import com.sjiwon.anotherart.global.exception.AnotherArtException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
