+++ backend/src/main/java/com/sjiwon/anotherart/art/domain/service/ArtRegistrationProcessor.java
@@ -2,9 +2,9 @@ package com.sjiwon.anotherart.art.domain.service;
 
 import com.sjiwon.anotherart.art.domain.model.Art;
 import com.sjiwon.anotherart.art.domain.repository.ArtRepository;
-import com.sjiwon.anotherart.auction.domain.Auction;
-import com.sjiwon.anotherart.auction.domain.AuctionRepository;
-import com.sjiwon.anotherart.auction.domain.Period;
+import com.sjiwon.anotherart.auction.domain.model.Auction;
+import com.sjiwon.anotherart.auction.domain.model.Period;
+import com.sjiwon.anotherart.auction.domain.repository.AuctionRepository;
 import com.sjiwon.anotherart.global.annotation.AnotherArtWritableTransactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
