+++ backend/src/main/java/com/sjiwon/anotherart/auction/application/usecase/BidService.java
@@ -1,6 +1,7 @@
-package com.sjiwon.anotherart.auction.service;
+package com.sjiwon.anotherart.auction.application.usecase;
 
-import com.sjiwon.anotherart.auction.domain.Auction;
+import com.sjiwon.anotherart.auction.domain.model.Auction;
+import com.sjiwon.anotherart.auction.domain.repository.AuctionRepository;
 import com.sjiwon.anotherart.member.domain.model.Member;
 import com.sjiwon.anotherart.member.domain.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
@@ -12,11 +13,11 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class BidService {
     private final MemberRepository memberRepository;
-    private final AuctionFindService auctionFindService;
+    private final AuctionRepository auctionRepository;
 
     @Transactional
     public void bid(final Long auctionId, final Long bidderId, final Integer bidPrice) {
-        final Auction auction = auctionFindService.findById(auctionId);
+        final Auction auction = auctionRepository.getById(auctionId);
         final Member bidder = memberRepository.getById(bidderId);
 
         auction.applyNewBid(bidder, bidPrice);
