+++ backend/src/main/java/com/sjiwon/anotherart/auction/domain/repository/AuctionRecordRepository.java
@@ -1,5 +1,6 @@
-package com.sjiwon.anotherart.auction.domain.record;
+package com.sjiwon.anotherart.auction.domain.repository;
 
+import com.sjiwon.anotherart.auction.domain.model.AuctionRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AuctionRecordRepository extends JpaRepository<AuctionRecord, Long> {
