+++ backend/src/main/java/com/sjiwon/anotherart/auction/presentation/dto/request/BidRequest.java
@@ -1,4 +1,4 @@
-package com.sjiwon.anotherart.auction.controller.dto.request;
+package com.sjiwon.anotherart.auction.presentation.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
