+++ backend/src/test/java/com/sjiwon/anotherart/art/application/usecase/RegisterArtUseCaseTest.java
@@ -7,8 +7,8 @@ import com.sjiwon.anotherart.art.domain.service.ArtImageUploader;
 import com.sjiwon.anotherart.art.domain.service.ArtRegistrationProcessor;
 import com.sjiwon.anotherart.art.domain.service.ArtResourceValidator;
 import com.sjiwon.anotherart.art.exception.ArtErrorCode;
-import com.sjiwon.anotherart.auction.domain.Auction;
-import com.sjiwon.anotherart.auction.domain.AuctionRepository;
+import com.sjiwon.anotherart.auction.domain.model.Auction;
+import com.sjiwon.anotherart.auction.domain.repository.AuctionRepository;
 import com.sjiwon.anotherart.common.UseCaseTest;
 import com.sjiwon.anotherart.file.domain.model.RawFileData;
 import com.sjiwon.anotherart.file.utils.converter.FileConverter;
