+++ backend/src/test/java/com/sjiwon/anotherart/art/domain/repository/query/ArtSingleQueryRepositoryTest.java
@@ -4,8 +4,8 @@
 //import com.sjiwon.anotherart.art.domain.repository.ArtRepository;
 //import com.sjiwon.anotherart.art.infra.query.dto.response.AuctionArt;
 //import com.sjiwon.anotherart.art.infra.query.dto.response.GeneralArt;
-//import com.sjiwon.anotherart.auction.domain.Auction;
-//import com.sjiwon.anotherart.auction.domain.AuctionRepository;
+//import com.sjiwon.anotherart.auction.domain.model.Auction;
+//import com.sjiwon.anotherart.auction.domain.repository.AuctionRepository;
 //import com.sjiwon.anotherart.common.RepositoryTest;
 //import com.sjiwon.anotherart.common.fixture.ArtFixture;
 //import com.sjiwon.anotherart.common.fixture.MemberFixture;
