+++ backend/src/test/java/com/sjiwon/anotherart/auction/application/usecase/BidServiceConcurrencyTest.java
@@ -1,8 +1,8 @@
-package com.sjiwon.anotherart.auction.service;
+package com.sjiwon.anotherart.auction.application.usecase;
 
 import com.sjiwon.anotherart.art.domain.model.Art;
-import com.sjiwon.anotherart.auction.domain.Auction;
-import com.sjiwon.anotherart.auction.facade.BidFacade;
+import com.sjiwon.anotherart.auction.application.facade.BidFacade;
+import com.sjiwon.anotherart.auction.domain.model.Auction;
 import com.sjiwon.anotherart.common.ConcurrencyTest;
 import com.sjiwon.anotherart.member.domain.model.Member;
 import lombok.extern.slf4j.Slf4j;
