+++ backend/src/test/java/com/sjiwon/anotherart/auction/application/usecase/BidServiceTest.java
@@ -1,8 +1,8 @@
-package com.sjiwon.anotherart.auction.service;
+package com.sjiwon.anotherart.auction.application.usecase;
 
 import com.sjiwon.anotherart.art.domain.model.Art;
-import com.sjiwon.anotherart.auction.domain.Auction;
-import com.sjiwon.anotherart.auction.domain.record.AuctionRecord;
+import com.sjiwon.anotherart.auction.domain.model.Auction;
+import com.sjiwon.anotherart.auction.domain.model.AuctionRecord;
 import com.sjiwon.anotherart.auction.exception.AuctionErrorCode;
 import com.sjiwon.anotherart.common.ServiceTest;
 import com.sjiwon.anotherart.common.fixture.MemberFixture;
