+++ backend/src/test/java/com/sjiwon/anotherart/common/fixture/AuctionFixture.java
@@ -1,7 +1,7 @@
 package com.sjiwon.anotherart.common.fixture;
 
 import com.sjiwon.anotherart.art.domain.model.Art;
-import com.sjiwon.anotherart.auction.domain.Auction;
+import com.sjiwon.anotherart.auction.domain.model.Auction;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
