+++ backend/src/test/java/com/sjiwon/anotherart/purchase/domain/PurchaseTest.java
@@ -1,7 +1,7 @@
 package com.sjiwon.anotherart.purchase.domain;
 
 import com.sjiwon.anotherart.art.domain.model.Art;
-import com.sjiwon.anotherart.auction.domain.Auction;
+import com.sjiwon.anotherart.auction.domain.model.Auction;
 import com.sjiwon.anotherart.global.exception.AnotherArtException;
 import com.sjiwon.anotherart.member.domain.model.Member;
 import com.sjiwon.anotherart.member.exception.MemberErrorCode;
